+++ crates/debugger/src/lib.rs
@@ -7,10 +7,7 @@
 #[macro_use]
 extern crate tracing;
 
-mod builder;
-pub use builder::DebuggerBuilder;
-
 mod op;
 
 mod tui;
-pub use tui::{Debugger, ExitReason};
+pub use tui::{Debugger, DebuggerBuilder, ExitReason};
