+++ crates/debugger/src/tui/draw.rs
@@ -1,6 +1,6 @@
-//! The TUI implementation.
+//! TUI draw implementation.
 
-use crate::{op::OpcodeParam, DebuggerBuilder};
+use crate::{op::OpcodeParam, Debugger, DebuggerBuilder};
 use alloy_primitives::{Address, U256};
 use crossterm::{
     event::{
@@ -36,509 +36,12 @@ use std::{
     time::{Duration, Instant},
 };
 
-/// Used to indicate why the debugger quit.
-#[derive(Debug)]
-pub enum ExitReason {
-    /// Exit using 'q'.
-    CharExit,
-}
-
-/// The TUI debugger.
-pub struct Debugger {
-    debug_arena: Vec<(Address, Vec<DebugStep>, CallKind)>,
-    terminal: Terminal<CrosstermBackend<io::Stdout>>,
-    identified_contracts: HashMap<Address, String>,
-    /// Source map of contract sources
-    contracts_sources: ContractSources,
-    /// A mapping of source -> (PC -> IC map for deploy code, PC -> IC map for runtime code)
-    pc_ic_maps: BTreeMap<String, (PCICMap, PCICMap)>,
-    breakpoints: Breakpoints,
-}
+use super::context::DrawMemory;
 
 impl Debugger {
-    /// Creates a new debugger builder.
-    #[inline]
-    pub fn builder() -> DebuggerBuilder {
-        DebuggerBuilder::new()
-    }
-
-    /// Creates a new debugger.
-    pub fn new(
-        debug_arena: Vec<(Address, Vec<DebugStep>, CallKind)>,
-        identified_contracts: HashMap<Address, String>,
-        contracts_sources: ContractSources,
-        breakpoints: Breakpoints,
-    ) -> Result<Self> {
-        let backend = CrosstermBackend::new(io::stdout());
-        let terminal = Terminal::new(backend)?;
-        let pc_ic_maps = contracts_sources
-            .0
-            .iter()
-            .flat_map(|(contract_name, files_sources)| {
-                files_sources.iter().filter_map(|(_, (_, contract))| {
-                    Some((
-                        contract_name.clone(),
-                        (
-                            build_pc_ic_map(
-                                SpecId::LATEST,
-                                contract.bytecode.object.as_bytes()?.as_ref(),
-                            ),
-                            build_pc_ic_map(
-                                SpecId::LATEST,
-                                contract
-                                    .deployed_bytecode
-                                    .bytecode
-                                    .as_ref()?
-                                    .object
-                                    .as_bytes()?
-                                    .as_ref(),
-                            ),
-                        ),
-                    ))
-                })
-            })
-            .collect();
-        Ok(Self {
-            debug_arena,
-            terminal,
-            identified_contracts,
-            contracts_sources,
-            pc_ic_maps,
-            breakpoints,
-        })
-    }
-
-    /// Starts the debugger TUI. Terminates the current process on failure or user exit.
-    pub fn run_exit(mut self) -> ! {
-        let code = match self.try_run() {
-            Ok(ExitReason::CharExit) => 0,
-            Err(e) => {
-                println!("{e}");
-                1
-            }
-        };
-        std::process::exit(code)
-    }
-
-    /// Starts the debugger TUI.
-    pub fn try_run(&mut self) -> Result<ExitReason> {
-        let mut guard = DebuggerGuard::setup(self)?;
-        let r = guard.0.try_run_real();
-        // Cleanup only once.
-        guard.restore()?;
-        std::mem::forget(guard);
-        r
-    }
-
-    #[instrument(target = "debugger", name = "run", skip_all, ret)]
-    fn try_run_real(&mut self) -> Result<ExitReason> {
-        // Create the context.
-        let mut cx = DebuggerContext::new(self);
-        cx.init()?;
-
-        // Create an event listener in a different thread.
-        let (tx, rx) = mpsc::channel();
-        thread::Builder::new()
-            .name("event-listener".into())
-            .spawn(move || Self::event_listener(tx))
-            .expect("failed to spawn thread");
-
-        loop {
-            match cx.handle_event(rx.recv()?) {
-                ControlFlow::Continue(()) => {}
-                ControlFlow::Break(reason) => return Ok(reason),
-            }
-            cx.draw()?;
-        }
-    }
-
-    fn event_listener(tx: mpsc::Sender<Event>) {
-        // This is the recommend tick rate from `ratatui`, based on their examples
-        let tick_rate = Duration::from_millis(200);
-
-        let mut last_tick = Instant::now();
-        loop {
-            // Poll events since last tick - if last tick is greater than tick_rate, we
-            // demand immediate availability of the event. This may affect interactivity,
-            // but I'm not sure as it is hard to test.
-            if event::poll(tick_rate.saturating_sub(last_tick.elapsed())).unwrap() {
-                let event = event::read().unwrap();
-                if tx.send(event).is_err() {
-                    return;
-                }
-            }
-
-            // Force update if time has passed
-            if last_tick.elapsed() > tick_rate {
-                last_tick = Instant::now();
-            }
-        }
-    }
-}
-
-/// This is currently used to remember last scroll position so screen doesn't wiggle as much.
-#[derive(Default)]
-struct DrawMemory {
-    current_startline: usize,
-    inner_call_index: usize,
-    current_mem_startline: usize,
-    current_stack_startline: usize,
-}
-
-/// Handles terminal state. `restore` should be called before drop to handle errors.
-#[must_use]
-struct DebuggerGuard<'a>(&'a mut Debugger);
-
-impl<'a> DebuggerGuard<'a> {
-    fn setup(dbg: &'a mut Debugger) -> Result<Self> {
-        let this = Self(dbg);
-        enable_raw_mode()?;
-        execute!(*this.0.terminal.backend_mut(), EnterAlternateScreen, EnableMouseCapture)?;
-        this.0.terminal.hide_cursor()?;
-        Ok(this)
-    }
-
-    fn restore(&mut self) -> Result<()> {
-        disable_raw_mode()?;
-        execute!(*self.0.terminal.backend_mut(), LeaveAlternateScreen, DisableMouseCapture)?;
-        self.0.terminal.show_cursor()?;
-        Ok(())
-    }
-}
-
-impl Drop for DebuggerGuard<'_> {
-    #[inline]
-    fn drop(&mut self) {
-        let _ = self.restore();
-    }
-}
-
-struct DebuggerContext<'a> {
-    debugger: &'a mut Debugger,
-
-    /// Buffer for keys prior to execution, i.e. '10' + 'k' => move up 10 operations.
-    key_buffer: String,
-    /// Current step in the debug steps.
-    current_step: usize,
-    draw_memory: DrawMemory,
-    opcode_list: Vec<String>,
-    last_index: usize,
-
-    stack_labels: bool,
-    mem_utf: bool,
-    show_shortcuts: bool,
-}
-
-impl<'a> DebuggerContext<'a> {
-    fn new(debugger: &'a mut Debugger) -> Self {
-        DebuggerContext {
-            debugger,
-
-            key_buffer: String::with_capacity(64),
-            current_step: 0,
-            draw_memory: DrawMemory::default(),
-            opcode_list: Vec::new(),
-            last_index: 0,
-
-            stack_labels: false,
-            mem_utf: false,
-            show_shortcuts: true,
-        }
-    }
-
-    fn init(&mut self) -> Result<()> {
-        self.debugger.terminal.clear()?;
-        self.gen_opcode_list();
-        Ok(())
-    }
-
-    fn debug_arena(&self) -> &[(Address, Vec<DebugStep>, CallKind)] {
-        &self.debugger.debug_arena
-    }
-
-    fn gen_opcode_list(&mut self) {
-        self.opcode_list = self.opcode_list();
-    }
-
-    fn opcode_list(&self) -> Vec<String> {
-        self.debugger.debug_arena[self.draw_memory.inner_call_index]
-            .1
-            .iter()
-            .map(DebugStep::pretty_opcode)
-            .collect()
-    }
-}
-
-impl DebuggerContext<'_> {
-    fn handle_event(&mut self, event: Event) -> ControlFlow<ExitReason> {
-        if self.last_index != self.draw_memory.inner_call_index {
-            self.opcode_list = self.debug_arena()[self.draw_memory.inner_call_index]
-                .1
-                .iter()
-                .map(|step| step.pretty_opcode())
-                .collect();
-            self.last_index = self.draw_memory.inner_call_index;
-        }
-
-        match event {
-            Event::Key(event) => self.handle_key_event(event),
-            Event::Mouse(event) => self.handle_mouse_event(event),
-            _ => ControlFlow::Continue(()),
-        }
-    }
-
-    fn handle_key_event(&mut self, event: KeyEvent) -> ControlFlow<ExitReason> {
-        if let KeyCode::Char(c) = event.code {
-            if c.is_alphanumeric() || c == '\'' {
-                self.handle_breakpoint(c);
-                return ControlFlow::Continue(());
-            }
-        }
-
-        match event.code {
-            // Exit
-            KeyCode::Char('q') => return ControlFlow::Break(ExitReason::CharExit),
-            // Move down
-            KeyCode::Char('j') | KeyCode::Down => {
-                // Grab number of times to do it
-                for _ in 0..Debugger::buffer_as_number(&self.key_buffer, 1) {
-                    if event.modifiers.contains(KeyModifiers::CONTROL) {
-                        let max_mem = (self.debug_arena()[self.draw_memory.inner_call_index].1
-                            [self.current_step]
-                            .memory
-                            .len() /
-                            32)
-                        .saturating_sub(1);
-                        if self.draw_memory.current_mem_startline < max_mem {
-                            self.draw_memory.current_mem_startline += 1;
-                        }
-                    } else if self.current_step < self.opcode_list.len() - 1 {
-                        self.current_step += 1;
-                    } else if self.draw_memory.inner_call_index < self.debug_arena().len() - 1 {
-                        self.draw_memory.inner_call_index += 1;
-                        self.current_step = 0;
-                    }
-                }
-                self.key_buffer.clear();
-            }
-            KeyCode::Char('J') => {
-                for _ in 0..Debugger::buffer_as_number(&self.key_buffer, 1) {
-                    let max_stack = self.debug_arena()[self.draw_memory.inner_call_index].1
-                        [self.current_step]
-                        .stack
-                        .len()
-                        .saturating_sub(1);
-                    if self.draw_memory.current_stack_startline < max_stack {
-                        self.draw_memory.current_stack_startline += 1;
-                    }
-                }
-                self.key_buffer.clear();
-            }
-            // Move up
-            KeyCode::Char('k') | KeyCode::Up => {
-                for _ in 0..Debugger::buffer_as_number(&self.key_buffer, 1) {
-                    if event.modifiers.contains(KeyModifiers::CONTROL) {
-                        self.draw_memory.current_mem_startline =
-                            self.draw_memory.current_mem_startline.saturating_sub(1);
-                    } else if self.current_step > 0 {
-                        self.current_step -= 1;
-                    } else if self.draw_memory.inner_call_index > 0 {
-                        self.draw_memory.inner_call_index -= 1;
-                        self.current_step =
-                            self.debug_arena()[self.draw_memory.inner_call_index].1.len() - 1;
-                    }
-                }
-                self.key_buffer.clear();
-            }
-            KeyCode::Char('K') => {
-                for _ in 0..Debugger::buffer_as_number(&self.key_buffer, 1) {
-                    self.draw_memory.current_stack_startline =
-                        self.draw_memory.current_stack_startline.saturating_sub(1);
-                }
-                self.key_buffer.clear();
-            }
-            // Go to top of file
-            KeyCode::Char('g') => {
-                self.draw_memory.inner_call_index = 0;
-                self.current_step = 0;
-                self.key_buffer.clear();
-            }
-            // Go to bottom of file
-            KeyCode::Char('G') => {
-                self.draw_memory.inner_call_index = self.debug_arena().len() - 1;
-                self.current_step =
-                    self.debug_arena()[self.draw_memory.inner_call_index].1.len() - 1;
-                self.key_buffer.clear();
-            }
-            // Go to previous call
-            KeyCode::Char('c') => {
-                self.draw_memory.inner_call_index =
-                    self.draw_memory.inner_call_index.saturating_sub(1);
-                self.current_step =
-                    self.debug_arena()[self.draw_memory.inner_call_index].1.len() - 1;
-                self.key_buffer.clear();
-            }
-            // Go to next call
-            KeyCode::Char('C') => {
-                if self.debug_arena().len() > self.draw_memory.inner_call_index + 1 {
-                    self.draw_memory.inner_call_index += 1;
-                    self.current_step = 0;
-                }
-                self.key_buffer.clear();
-            }
-            // Step forward
-            KeyCode::Char('s') => {
-                for _ in 0..Debugger::buffer_as_number(&self.key_buffer, 1) {
-                    let remaining_ops = self.opcode_list[self.current_step..].to_vec();
-                    self.current_step += remaining_ops
-                        .iter()
-                        .enumerate()
-                        .find_map(|(i, op)| {
-                            if i < remaining_ops.len() - 1 {
-                                match (
-                                    op.contains("JUMP") && op != "JUMPDEST",
-                                    &*remaining_ops[i + 1],
-                                ) {
-                                    (true, "JUMPDEST") => Some(i + 1),
-                                    _ => None,
-                                }
-                            } else {
-                                None
-                            }
-                        })
-                        .unwrap_or(self.opcode_list.len() - 1);
-                    if self.current_step > self.opcode_list.len() {
-                        self.current_step = self.opcode_list.len() - 1
-                    };
-                }
-                self.key_buffer.clear();
-            }
-            // Step backwards
-            KeyCode::Char('a') => {
-                for _ in 0..Debugger::buffer_as_number(&self.key_buffer, 1) {
-                    let prev_ops = self.opcode_list[..self.current_step].to_vec();
-                    self.current_step = prev_ops
-                        .iter()
-                        .enumerate()
-                        .rev()
-                        .find_map(|(i, op)| {
-                            if i > 0 {
-                                match (
-                                    prev_ops[i - 1].contains("JUMP") &&
-                                        prev_ops[i - 1] != "JUMPDEST",
-                                    &**op,
-                                ) {
-                                    (true, "JUMPDEST") => Some(i - 1),
-                                    _ => None,
-                                }
-                            } else {
-                                None
-                            }
-                        })
-                        .unwrap_or_default();
-                }
-                self.key_buffer.clear();
-            }
-            // toggle stack labels
-            KeyCode::Char('t') => self.stack_labels = !self.stack_labels,
-            // toggle memory utf8 decoding
-            KeyCode::Char('m') => self.mem_utf = !self.mem_utf,
-            // toggle help notice
-            KeyCode::Char('h') => self.show_shortcuts = !self.show_shortcuts,
-            KeyCode::Char(other) => match other {
-                '0' | '1' | '2' | '3' | '4' | '5' | '6' | '7' | '8' | '9' | '\'' => {
-                    self.key_buffer.push(other);
-                }
-                // Invalid key, clear buffer
-                _ => self.key_buffer.clear(),
-            },
-            _ => self.key_buffer.clear(),
-        };
-
-        ControlFlow::Continue(())
-    }
-
-    fn handle_breakpoint(&mut self, c: char) {
-        // Find the location of the called breakpoint in the whole debug arena (at this address with
-        // this pc)
-        if let Some((caller, pc)) = self.debugger.breakpoints.get(&c) {
-            for (i, (_caller, debug_steps, _)) in self.debug_arena().iter().enumerate() {
-                if _caller == caller {
-                    if let Some(step) = debug_steps.iter().position(|step| step.pc == *pc) {
-                        self.draw_memory.inner_call_index = i;
-                        self.current_step = step;
-                        break
-                    }
-                }
-            }
-        }
-        self.key_buffer.clear();
-    }
-
-    fn handle_mouse_event(&mut self, event: MouseEvent) -> ControlFlow<ExitReason> {
-        match event.kind {
-            MouseEventKind::ScrollUp => {
-                if self.current_step > 0 {
-                    self.current_step -= 1;
-                } else if self.draw_memory.inner_call_index > 0 {
-                    self.draw_memory.inner_call_index -= 1;
-                    self.draw_memory.current_mem_startline = 0;
-                    self.draw_memory.current_stack_startline = 0;
-                    self.current_step =
-                        self.debug_arena()[self.draw_memory.inner_call_index].1.len() - 1;
-                }
-            }
-            MouseEventKind::ScrollDown => {
-                if self.current_step < self.opcode_list.len() - 1 {
-                    self.current_step += 1;
-                } else if self.draw_memory.inner_call_index < self.debug_arena().len() - 1 {
-                    self.draw_memory.inner_call_index += 1;
-                    self.draw_memory.current_mem_startline = 0;
-                    self.draw_memory.current_stack_startline = 0;
-                    self.current_step = 0;
-                }
-            }
-            _ => {}
-        }
-
-        ControlFlow::Continue(())
-    }
-
-    fn draw(&mut self) -> Result<()> {
-        self.debugger.terminal.draw(|f| {
-            let debug_arena = &self.debugger.debug_arena;
-            Debugger::draw_layout(
-                f,
-                debug_arena[self.draw_memory.inner_call_index].0,
-                &self.debugger.identified_contracts,
-                &self.debugger.pc_ic_maps,
-                &self.debugger.contracts_sources,
-                &debug_arena[self.draw_memory.inner_call_index].1[..],
-                &self.opcode_list,
-                self.current_step,
-                debug_arena[self.draw_memory.inner_call_index].2,
-                &mut self.draw_memory,
-                self.stack_labels,
-                self.mem_utf,
-                self.show_shortcuts,
-            )
-        })?;
-        Ok(())
-    }
-}
-
-impl Debugger {
-    /// Grab number from buffer. Used for something like '10k' to move up 10 operations
-    fn buffer_as_number(s: &str, default_value: usize) -> usize {
-        match s.parse() {
-            Ok(num) if num >= 1 => num,
-            _ => default_value,
-        }
-    }
-
     /// Create layout and subcomponents
     #[allow(clippy::too_many_arguments)]
-    fn draw_layout(
+    pub(crate) fn draw_layout(
         f: &mut Frame<'_>,
         address: Address,
         identified_contracts: &HashMap<Address, String>,
